sentence change
Changed the sentence: 'The user keys in global latitude and longitude coordinates before pressing a Get Weather button to view information about the weather at that location.'
Edited to: ' The user enters global latitude and longitude coordinates before pressing the Get Weather button to display the weather at that location.
I feel that the original sentence was a little too difficult to understand. The new sentence is clearer.